+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Added links to related repos in README.
 
+### Changed
+- Streamlined `SplitMeshDef` usage for `MergedMesh.split_mesh()` property passthrough.
+
 ## [2.4.0] - 2026-09-06
 
 ### Added